Rocket movement is now centered on image center, and not corner

+++ src/Model/help/DegreesCalculations.h
@@ -11,9 +11,9 @@
 
 class DegreesCalculations {
 public:
-	static Rotation radiansToDegrees( float32 rad);
+	static Rotation radiansToDegrees( double rad);
 
-	static float degreesToRadians( Rotation degrees );
+	static double degreesToRadians( Rotation degrees );
 };
